20150713 init base

+++ src/java/com/android/internal/telephony/InboundSmsHandler.java
@@ -1261,7 +1261,7 @@ public abstract class InboundSmsHandler extends StateMachine {
         if (cr.acquireProvider(FIREWALL_PROVIDER_URI) != null && null != address) {
             Bundle extras = new Bundle();
             extras.putString(EXTRA_NUMBER, address);
-            extras.putInt(PhoneConstants.SUBSCRIPTION_KEY, mPhone.getSubId());
+            extras.putLong(PhoneConstants.SUBSCRIPTION_KEY, mPhone.getSubId());
             extras = cr.call(FIREWALL_PROVIDER_URI, IS_FORBIDDEN, null, extras);
             if (extras != null) {
                 isForbidden = extras.getBoolean(IS_FORBIDDEN);
